server/requirements.txt
fastapi[standard]
python-dotenv
uvicorn
pydantic
rich
python-multipart
tensorflow==2.13.0
numpy
pillow
vit-keras
tensorflow-addons
typing-extensions
opencv-python